requirements.txt
anthropic==0.71.0
fastapi==0.109.0
pydantic==2.5.3
python-dotenv==1.0.0
uvicorn==0.25.0
